move file watcher tests in with the other hot reload tests
